feat: :sparkles: static import full-calendar and rearrange files to make svelte component
