+++ src/routes/calendar/+page.svelte
@@ -1,29 +1,8 @@
 <script lang="ts">
-	import type { Calendar } from '@fullcalendar/core/index.js';
-	import { onDestroy, onMount } from 'svelte';
-
-	let calendarContainer: HTMLDivElement;
-	let calendarToDestroy: Calendar;
-
-	onMount(async () => {
-		const { generateCalendar } = await import('./full-calendar');
-		calendarToDestroy = generateCalendar(calendarContainer);
-	});
-	onDestroy(() => {
-		calendarToDestroy.destroy();
-	});
+	import FullCalendar from '$lib/FullCalendar.svelte';
 </script>
 
-<div id="calendar-container" bind:this={calendarContainer} />
-
-<style>
-	:global(main:has(#calendar-container)) {
-		max-width: none;
-	}
-
-	#calendar-container {
-		margin: 0 auto;
-		font-size: min(18px, 0.75em);
-		max-width: max(640px, calc((4 / 3) * (100lvh - 200px)));
-	}
-</style>
+<FullCalendar
+	googleCalendarApiKey="AIzaSyDHUqblOxNndbN7jUXvoKy8IugKLlSXbkE"
+	googleCalendarId="c_1pcp6odi9qfe276tpuob8h00ms@group.calendar.google.com"
+/>
